+++ packages/core-js-compat/src/data.js
@@ -1331,6 +1331,8 @@ const data = {
   },
   'esnext.symbol.pattern-match': {
   },
+  'esnext.symbol.replace-all': {
+  },
   'esnext.weak-map.delete-all': {
   },
   'esnext.weak-map.from': {