+++ packages/core-js/features/symbol/index.js
@@ -3,3 +3,4 @@ module.exports = require('../../es/symbol');
 require('../../modules/esnext.symbol.dispose');
 require('../../modules/esnext.symbol.observable');
 require('../../modules/esnext.symbol.pattern-match');
+require('../../modules/esnext.symbol.replace-all');