+++ packages/core-js/features/symbol/replace-all.js
@@ -1,0 +1,3 @@
+require('../../modules/esnext.symbol.replace-all');
+
+module.exports = require('../../internals/wrapped-well-known-symbol').f('replaceAll');