+++ packages/core-js/proposals/string-replace-all.js
@@ -1,1 +1,2 @@
 require('../modules/esnext.string.replace-all');
+require('../modules/esnext.symbol.replace-all');