+++ packages/core-js/stage/2.js
@@ -1,3 +1,4 @@
 require('../proposals/set-methods');
+require('../proposals/string-replace-all');
 
 module.exports = require('./3');